Extends the 'ipfs add' options
In preparation for extending the options that the add command will accept the tests for add, cat and get are pulled into their own test files. The previous implementation of the -q quiet flag has be updated to have the same behaviour as the --quiet flag in go-ipfs
